requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
Cython==0.29.34
distlib==0.3.6
django-os-geocoder==0.1.6
docutils==0.20.1
filelock==3.12.0
gps==3.19
idna==3.4
Kivy==2.2.0rc1
Kivy-Garden==0.1.5
materialyoucolor==1.2.0
Pillow==9.5.0
platformdirs==3.5.1
plyer==2.1.0
progressbar==2.5
Pygments==2.15.1
python-dotenv==1.0.0
requests==2.30.0
South==1.0.2
urllib3==2.0.2
virtualenv==20.23.0
watchdog==3.0.0
